fix(clippy): remove redundant cast and single-char push_str in Tier B PTX emitters
Five Clippy violations (-D warnings) introduced by the PCA Tier B.2
predicate work (PR #169):

- s_compute.rs:94: block_kv is already u32 (assigned line 39); the
  additional `as u32` cast is a no-op. Removed.
- pca_tilerange.rs:187,209,457,596: push_str("\n") changed to push('\n')
  per clippy::single_char_add_str.

No behavioural change — lint fixes only. cargo clippy --workspace
-- -D warnings clean; 1776/1776 nsl-codegen lib tests pass.

+++ crates/nsl-codegen/src/flash_attention_v2/phases/forward/s_compute.rs
@@ -91,7 +91,7 @@ pub fn emit(
     //     Codified in `docs/superpowers/specs/2026-05-13-tier-b-b15-3-skip-ratio-investigation.md`.
     if let Some((seq_len, residency)) = tier_b {
         if crate::pca_tilerange::should_emit_tier_b(config, seq_len as u64, residency) {
-            let log2_bkv = (block_kv as u32).trailing_zeros();
+            let log2_bkv = block_kv.trailing_zeros();
             let range_table_base =
                 crate::flash_attention_v2::smem_layout::tier_b_range_table_offset(
                     config,
